+++ src/main/webapp/wise5/components/outsideURL/outsideURLAuthoringController.js
@@ -1,0 +1,294 @@
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+
+var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
+
+var _outsideURLController = require('./outsideURLController');
+
+var _outsideURLController2 = _interopRequireDefault(_outsideURLController);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
+
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+
+var OutsideURLAuthoringController = function (_OutsideURLController) {
+  _inherits(OutsideURLAuthoringController, _OutsideURLController);
+
+  function OutsideURLAuthoringController($filter, $mdDialog, $q, $rootScope, $sce, $scope, AnnotationService, ConfigService, NodeService, NotebookService, OutsideURLService, ProjectService, StudentAssetService, StudentDataService, UtilService) {
+    _classCallCheck(this, OutsideURLAuthoringController);
+
+    // generate the summernote rubric element id
+    var _this = _possibleConstructorReturn(this, (OutsideURLAuthoringController.__proto__ || Object.getPrototypeOf(OutsideURLAuthoringController)).call(this, $filter, $mdDialog, $q, $rootScope, $sce, $scope, AnnotationService, ConfigService, NodeService, NotebookService, OutsideURLService, ProjectService, StudentAssetService, StudentDataService, UtilService));
+
+    _this.summernoteRubricId = 'summernoteRubric_' + _this.nodeId + '_' + _this.componentId;
+
+    // set the component rubric into the summernote rubric
+    _this.summernoteRubricHTML = _this.componentContent.rubric;
+
+    // the tooltip text for the insert WISE asset button
+    var insertAssetString = _this.$translate('html.insertAsset');
+
+    /*
+     * create the custom button for inserting WISE assets into
+     * summernote
+     */
+    var InsertAssetButton = _this.UtilService.createInsertAssetButton(_this, null, _this.nodeId, _this.componentId, 'rubric', insertAssetString);
+
+    /*
+     * the options that specifies the tools to display in the
+     * summernote prompt
+     */
+    _this.summernoteRubricOptions = {
+      toolbar: [['style', ['style']], ['font', ['bold', 'underline', 'clear']], ['fontname', ['fontname']], ['fontsize', ['fontsize']], ['color', ['color']], ['para', ['ul', 'ol', 'paragraph']], ['table', ['table']], ['insert', ['link', 'video']], ['view', ['fullscreen', 'codeview', 'help']], ['customButton', ['insertAssetButton']]],
+      height: 300,
+      disableDragAndDrop: true,
+      buttons: {
+        insertAssetButton: InsertAssetButton
+      }
+    };
+
+    _this.updateAdvancedAuthoringView();
+
+    $scope.$watch(function () {
+      return _this.authoringComponentContent;
+    }, function (newValue, oldValue) {
+      _this.componentContent = _this.ProjectService.injectAssetPaths(newValue);
+
+      // set the url
+      _this.setURL(_this.authoringComponentContent.url);
+    }, true);
+
+    /*
+     * Listen for the assetSelected event which occurs when the user
+     * selects an asset from the choose asset popup
+     */
+    _this.$scope.$on('assetSelected', function (event, args) {
+
+      if (args != null) {
+
+        // make sure the event was fired for this component
+        if (args.nodeId == _this.nodeId && args.componentId == _this.componentId) {
+          // the asset was selected for this component
+          var assetItem = args.assetItem;
+
+          if (assetItem != null) {
+            var fileName = assetItem.fileName;
+
+            if (fileName != null) {
+              /*
+               * get the assets directory path
+               * e.g.
+               * /wise/curriculum/3/
+               */
+              var assetsDirectoryPath = _this.ConfigService.getProjectAssetsDirectoryPath();
+              var fullAssetPath = assetsDirectoryPath + '/' + fileName;
+
+              var summernoteId = '';
+
+              if (args.target == 'prompt') {
+                // the target is the summernote prompt element
+                summernoteId = 'summernotePrompt_' + _this.nodeId + '_' + _this.componentId;
+              } else if (args.target == 'rubric') {
+                // the target is the summernote rubric element
+                summernoteId = 'summernoteRubric_' + _this.nodeId + '_' + _this.componentId;
+              }
+
+              if (summernoteId != '') {
+                if (_this.UtilService.isImage(fileName)) {
+                  /*
+                   * move the cursor back to its position when the asset chooser
+                   * popup was clicked
+                   */
+                  $('#' + summernoteId).summernote('editor.restoreRange');
+                  $('#' + summernoteId).summernote('editor.focus');
+
+                  // add the image html
+                  $('#' + summernoteId).summernote('insertImage', fullAssetPath, fileName);
+                } else if (_this.UtilService.isVideo(fileName)) {
+                  /*
+                   * move the cursor back to its position when the asset chooser
+                   * popup was clicked
+                   */
+                  $('#' + summernoteId).summernote('editor.restoreRange');
+                  $('#' + summernoteId).summernote('editor.focus');
+
+                  // insert the video element
+                  var videoElement = document.createElement('video');
+                  videoElement.controls = 'true';
+                  videoElement.innerHTML = '<source ng-src="' + fullAssetPath + '" type="video/mp4">';
+                  $('#' + summernoteId).summernote('insertNode', videoElement);
+                }
+              }
+            }
+          }
+        }
+      }
+
+      // close the popup
+      _this.$mdDialog.hide();
+    });
+
+    /*
+     * The advanced button for a component was clicked. If the button was
+     * for this component, we will show the advanced authoring.
+     */
+    _this.$scope.$on('componentAdvancedButtonClicked', function (event, args) {
+      if (args != null) {
+        var componentId = args.componentId;
+        if (_this.componentId === componentId) {
+          _this.showAdvancedAuthoring = !_this.showAdvancedAuthoring;
+        }
+      }
+    });
+    return _this;
+  }
+
+  /**
+   * The component has changed in the regular authoring view so we will save the project
+   */
+
+
+  _createClass(OutsideURLAuthoringController, [{
+    key: 'authoringViewComponentChanged',
+    value: function authoringViewComponentChanged() {
+
+      // set the url
+      //this.setURL(this.authoringComponentContent.url);
+
+      // update the JSON string in the advanced authoring view textarea
+      this.updateAdvancedAuthoringView();
+
+      /*
+       * notify the parent node that the content has changed which will save
+       * the project to the server
+       */
+      this.$scope.$parent.nodeAuthoringController.authoringViewNodeChanged();
+    }
+  }, {
+    key: 'advancedAuthoringViewComponentChanged',
+
+
+    /**
+     * The component has changed in the advanced authoring view so we will update
+     * the component and save the project.
+     */
+    value: function advancedAuthoringViewComponentChanged() {
+
+      try {
+        /*
+         * create a new component by converting the JSON string in the advanced
+         * authoring view into a JSON object
+         */
+        var authoringComponentContent = angular.fromJson(this.authoringComponentContentJSONString);
+
+        // replace the component in the project
+        this.ProjectService.replaceComponent(this.nodeId, this.componentId, authoringComponentContent);
+
+        // set the new authoring component content
+        this.authoringComponentContent = authoringComponentContent;
+
+        // set the component content
+        this.componentContent = this.ProjectService.injectAssetPaths(authoringComponentContent);
+
+        /*
+         * notify the parent node that the content has changed which will save
+         * the project to the server
+         */
+        this.$scope.$parent.nodeAuthoringController.authoringViewNodeChanged();
+      } catch (e) {
+        this.$scope.$parent.nodeAuthoringController.showSaveErrorAdvancedAuthoring();
+      }
+    }
+  }, {
+    key: 'updateAdvancedAuthoringView',
+
+
+    /**
+     * Update the component JSON string that will be displayed in the advanced authoring view textarea
+     */
+    value: function updateAdvancedAuthoringView() {
+      this.authoringComponentContentJSONString = angular.toJson(this.authoringComponentContent, 4);
+    }
+  }, {
+    key: 'summernoteRubricHTMLChanged',
+
+
+    /**
+     * The author has changed the rubric
+     */
+    value: function summernoteRubricHTMLChanged() {
+
+      // get the summernote rubric html
+      var html = this.summernoteRubricHTML;
+
+      /*
+       * remove the absolute asset paths
+       * e.g.
+       * <img src='https://wise.berkeley.edu/curriculum/3/assets/sun.png'/>
+       * will be changed to
+       * <img src='sun.png'/>
+       */
+      html = this.ConfigService.removeAbsoluteAssetPaths(html);
+
+      /*
+       * replace <a> and <button> elements with <wiselink> elements when
+       * applicable
+       */
+      html = this.UtilService.insertWISELinks(html);
+
+      // update the component rubric
+      this.authoringComponentContent.rubric = html;
+
+      // the authoring component content has changed so we will save the project
+      this.authoringViewComponentChanged();
+    }
+
+    /**
+     * The show JSON button was clicked to show or hide the JSON authoring
+     */
+
+  }, {
+    key: 'showJSONButtonClicked',
+    value: function showJSONButtonClicked() {
+      // toggle the JSON authoring textarea
+      this.showJSONAuthoring = !this.showJSONAuthoring;
+
+      if (this.jsonStringChanged && !this.showJSONAuthoring) {
+        /*
+         * the author has changed the JSON and has just closed the JSON
+         * authoring view so we will save the component
+         */
+        this.advancedAuthoringViewComponentChanged();
+
+        // scroll to the top of the component
+        this.$rootScope.$broadcast('scrollToComponent', { componentId: this.componentId });
+
+        this.jsonStringChanged = false;
+      }
+    }
+
+    /**
+     * The author has changed the JSON manually in the advanced view
+     */
+
+  }, {
+    key: 'authoringJSONChanged',
+    value: function authoringJSONChanged() {
+      this.jsonStringChanged = true;
+    }
+  }]);
+
+  return OutsideURLAuthoringController;
+}(_outsideURLController2.default);
+
+OutsideURLAuthoringController.$inject = ['$filter', '$mdDialog', '$q', '$rootScope', '$sce', '$scope', 'AnnotationService', 'ConfigService', 'NodeService', 'NotebookService', 'OutsideURLService', 'ProjectService', 'StudentAssetService', 'StudentDataService', 'UtilService'];
+
+exports.default = OutsideURLAuthoringController;
+//# sourceMappingURL=outsideURLAuthoringController.js.map
